Small refactoring, fix possible error during starting agent (move api.Nodes.GetNodes() in try...catch block).

+++ Aims.IISAgent.Module/Pipes/NamedPipeClient.cs
@@ -193,7 +193,7 @@ namespace Aims.IISAgent.Module.Pipes
 				uint numBytesToRead = 0;
 				var result = ReadFile(_handle, ptr, length, ref numBytesToRead, 0);
 
-				var lastError = (int)GetLastError();//TODO why it is casted to int?
+				var lastError = (int)GetLastError();//TODO why it (uint) is casted to int?
 				if (!result && lastError != 0x000000EA) // ERROR_MORE_DATA
 					throw new Exception($"Error in ReadFile(...). ErrorCode:{lastError}");
